+++ academy-backend/src/main/java/uz/jurayev/academy/repository/AddressRepository.java
@@ -7,7 +7,7 @@ import java.util.Optional;
 
 public interface AddressRepository extends JpaRepository<Address, Long> {
 
-    Optional<Address> findById(Integer id);
+    Optional<Address> findById(Long id);
 
-    void deleteById(Integer id);
+    void deleteById(Long id);
 }
