+++ academy-backend/src/main/java/uz/jurayev/academy/rest/AddressDTO.java
@@ -12,7 +12,7 @@ import java.util.Set;
 @NoArgsConstructor
 public class AddressDTO {
 
-    private Integer districtId;
+    private Long districtId;
     private String details;
 
 }
